Working on #19 Just to make players visible when they shoot and when they are hit

+++ server/src/me/eddiep/ghost/server/game/Entity.java
@@ -121,7 +121,7 @@ public abstract class Entity {
     }
 
     public void setAlpha(float alpha) {
-        this.alpha = (int) (alpha * 255);
+        this.alpha = (byte) (alpha * 255);
     }
 
     public boolean isVisible() {
